==> 08_Word_Embeddings/06-Existierende Wortvektoren.ipynb <==
import sys, os
ON_COLAB = 'google.colab' in sys.modules

if ON_COLAB:
    os.system("test -f cc.de.300.bin || wget  https://dl.fbaipublicfiles.com/fasttext/vectors-crawl/cc.de.300.bin.gz && gunzip cc.de.300.bin.gz")

!pip install fasttext

import fasttext
cc_ft = fasttext.load_model("cc.de.300.bin")

cc_ft.get_nearest_neighbors("apfel", k=20)

cc_ft.get_nearest_neighbors("Apfel", k=20)

cc_ft.get_analogies("Frau", "Mann", "König")

cc_ft.get_analogies("Frau", "Mann", "Onkel")

# Speicher freigeben für Colab
del cc_ft

if ON_COLAB:
    os.system("test -f emb-glove-de-wiki.txt || wget -O emb-glove-de-wiki.txt https://int-emb-glove-de-wiki.s3.eu-central-1.amazonaws.com/vectors.txt")
    os.system("test -f emb-word2vec-de-wiki.txt || wget -O emb-word2vec-de-wiki.txt https://int-emb-word2vec-de-wiki.s3.eu-central-1.amazonaws.com/vectors.txt")
    os.system("test -f emb-fasttext-de-wiki.bin || wget -O emb-fasttext-de-wiki.bin https://s3.eu-central-1.amazonaws.com/int-emb-fasttext-de-wiki/20180917/model.bin")

!pip install "gensim>=4.0.0"

from gensim.models import KeyedVectors
emb_glove = KeyedVectors.load_word2vec_format("emb-glove-de-wiki.txt", no_header=True)

emb_glove.similar_by_word("apfel")

emb_glove.most_similar(positive=["frau", "könig"], negative=["mann"], topn=10)

emb_glove.most_similar(positive=["frau", "onkel"], negative=["mann"], topn=10)

# Speicher freigeben für Colab
del emb_glove

emb_w2v = KeyedVectors.load_word2vec_format("emb-word2vec-de-wiki.txt", no_header=True)

emb_w2v.similar_by_word("b'apfel'")

emb_w2v.similar_by_word("b'frau'")

emb_w2v.most_similar(positive=["b'frau'", "b'k\\xc3\\xb6nig'"], negative=["b'mann'"], topn=10)

emb_w2v.most_similar(positive=["b'frau'", "b'onkel'"], negative=["b'mann'"], topn=10)

# Speicher freigeben für Colab
del emb_w2v

emb_ft = fasttext.load_model("emb-fasttext-de-wiki.bin")

emb_ft.get_nearest_neighbors("apfel")

emb_ft.get_analogies("frau", "mann", "könig")

emb_ft.get_analogies("frau", "mann", "onkel")

emb_ft.get_nearest_neighbors("Apfel")

# Speicher freigeben für Colab
del emb_ft

if ON_COLAB:
    os.system("test -f GoogleNews-vectors-negative300.bin || wget https://figshare.com/ndownloader/files/10798046 && mv 10798046 GoogleNews-vectors-negative300.bin")

google_w2v = KeyedVectors.load_word2vec_format("GoogleNews-vectors-negative300.bin", binary=True)

google_w2v.similar_by_word("woman")

google_w2v.most_similar(positive=["woman", "king"], negative=["man"], topn=10)

google_w2v.most_similar(positive=["woman", "uncle"], negative=["man"], topn=10)
