==> 08_Word_Embeddings/07-Semantische_Transformation.ipynb <==
import sys, os
ON_COLAB = 'google.colab' in sys.modules

if ON_COLAB:
    os.system("test -f heise-articles-2020.db || wget  https://datanizing.com/heiseacademy/nlp-course/blob/main/99_Common/heise-articles-2020.db.gz && gunzip heise-articles-2020.db.gz")
    newsticker_db = 'heise-articles-2020.db'
else:
    newsticker_db = '../99_Common/heise-articles-2020.db'

import sqlite3 
import pandas as pd

sql = sqlite3.connect(newsticker_db)
df = pd.read_sql("SELECT * FROM nlp_articles WHERE datePublished<'2021-01-01' ORDER BY datePublished", 
                 sql, index_col="id", parse_dates=["datePublished"])

from sklearn.feature_extraction.text import TfidfVectorizer
from spacy.lang.de.stop_words import STOP_WORDS as stop_words

tfidf_vectorizer = TfidfVectorizer(stop_words=list(stop_words), min_df=10, sublinear_tf=True, use_idf=False)
tfidf_vectors = tfidf_vectorizer.fit_transform(df["nav"])

if ON_COLAB:
    os.system("test -f heise-articles-2020.w2v || wget  https://datanizing.com/heiseacademy/nlp-course/blob/main/99_Common/heise-articles-2020.w2v.gz && gunzip heise-articles-2020.w2v.gz")
    w2v_file = "heise-articles-2020.w2v"
else:
    w2v_file = "../99_Common/heise-articles-2020.w2v"

!pip install "gensim>=4.0.0"

from gensim.models import KeyedVectors
w2v = KeyedVectors.load_word2vec_format(w2v_file)

from tqdm import tqdm
import numpy as np

# averaged word vectors
awv = []
fn = tfidf_vectorizer.get_feature_names_out()
for i in tqdm(range(tfidf_vectors.shape[0])):
    v = np.zeros(w2v.vector_size)
    # immer nur eine Zeile, rows wird nicht benötigt
    rows,cols = tfidf_vectors[i].nonzero()
    for c in cols:
        feature = fn[c]
        if feature in w2v.key_to_index :
            # TF/IDF als Gewicht des normierten Wortvektors
            wv = w2v[feature]
            v += tfidf_vectors[i][0, c] * wv / np.linalg.norm(wv)
    if np.linalg.norm(v) > 0:
        v = v/np.linalg.norm(v)
    awv.append(v)

from sklearn.metrics.pairwise import cosine_similarity

apple = w2v["apple"]

r = cosine_similarity(awv, [apple])

df.iloc[r.argmax()]

wv =  w2v["google"]/np.linalg.norm(w2v["google"])
wv += w2v["umsatz"]/np.linalg.norm(w2v["umsatz"])
r = cosine_similarity(awv, [wv])
df.iloc[r.argmax()]

wv =  w2v["google"]/np.linalg.norm(w2v["google"])
wv += w2v["umsatz"]/np.linalg.norm(w2v["umsatz"])
wv += w2v["microsoft"]/np.linalg.norm(w2v["microsoft"])
r = cosine_similarity(awv, [wv])
df.iloc[r.argmax()]

print(df.iloc[r.argmax()]["full_text"])
